+++ src/knx/qknxnetipdisconnectresponse.cpp
@@ -9,9 +9,8 @@
 
 QT_BEGIN_NAMESPACE
 
-QKnxNetIpDisconnectResponse::QKnxNetIpDisconnectResponse(quint8 channelId,
-        QKnxNetIp::Error status)
-    : QKnxNetIpFrame(quint16(QKnxNetIp::ServiceType::DisconnectResponse))
+QKnxNetIpDisconnectResponse::QKnxNetIpDisconnectResponse(quint8 channelId, QKnxNetIp::Error status)
+    : QKnxNetIpFrame(QKnxNetIp::ServiceType::DisconnectResponse)
 {
     QKnxNetIpPayload payload(channelId);
     payload.setByte(1, quint8(status));
